Completed load_one.launch.py launch file
I believe that every node in the package will be shutdown after rviz
is closed because I added on_exit=Shutdown() to the rviz node in the
launch file. ExecuteLocal is a parent of the ExecuteProcess object, which
is a parent of the Node object. The kwarg on_exit should
get passed all the way back to ExecuteLocal because node calls
super.__init__ and so does ExecuteProcess.