Completed task A.2, I think!
I added all .stl, .rviz, and .urdf.xacro files to the repository and
added an install function to the CMakeLists file as well. I also added
an install function for the launch directory. I added a find_package(rclcpp Required)
line as well, since the ROS documentation did that and I assume I'll need
that too eventually.

I'm not sure if my directory hierarchy is what matt wants, but I'm pretty
sure it'll work and I like it so that's how it's going to stay for now.

Modified urdf files to reference files within my package, and not
turtlebot3_description. I'll need to update the file extensions of those files
to .urdf.xacro instead of .urdf, but I will do that in nthe next commit
since I forgot this time.

Put the launchfile in declarative style as well.