Finished Task A.3
Moved rviz, stl, and urdf.xacro files from the /config directory. Had a
brain fart there. They belong in the top level directory of the package
in their own unique directories.

Added the diff_params.yaml file and use it in the urdf.xacro file for
the turtlebot to define parameters. Also added a conditional statement
to use different collision geometry for the base link based on what
the value of the collision-radius is.