Finished a lot of B.6, but not done yet.
There's something funky wiht my drawings, will have to figure it out in
the morning.

I had been doing the tranformations on twists so wrong, it took me a
while to figure out what i was acutally supposed to do. got stuck
for about an hour looking up how to do adjoint matrices in se(2), and
adjoint matrices for 3x3 matrices. Figured out that I could just
pretend everything was 3d for the sake of using kevin's easy equation
he has in his book.

the intput and output files for the implementation
are also finished and they are perfect.

Had to also go back and change how I calculated the angle from a given
rotation matrix, i was doing it in such a naiive way.