Task C.2 Complete
Created the nusim package, and added a publisher for the current time-
step and a service to reset the state of the node, which at this point
is just the current_timestep_ variable.

I need to look in to making the tests pass now, before it becomes a
crazy huge task later down the line.