Finished diff_drive, on to testing
In testing for FK and IK, trying to make these tests as good as possible.
I think I have all of the math implemented correctly, but it's hard to
know. I have to figure out some test cases by hand or in my head, and then
see if the implementations make sense.