Added launchfile and configured CMake and pkg.xml
Need to commit so i can switch to main and do the vcs thing in E.1.
Have configured the turtle_control node to accept all the parameters in
diff_params.yaml and set them as private variables in the node, and also
wrote a launch file to grab that param file and use it to fill the node
with parameters.