Red robot is moving fine
Think I am about to have everything in place to allow the blue robot
to mirror the red robot in simulation before adding noise and everything.

Got stuck forever last night trying to figure out why the robot was
accelerating infinitely, but figured it out in 15 min when I woke up this
morning.