After much turmoil, it's works the same way
I get home from the lab and learn that my simulation is broken after
getting the real robot to work, so I spent the last two hours figuring
it out. I moved the sensor_data publisher to the timer callback inside
nusim, and that seemed to fix everything. Glad that FK or IK wasn't
broken.