The robot moves at the correct speed linearly in sim
Many issues with getting the robot to move at the correct speed in the
sim. At least at this point the red robot is behaving correctly. The
odometry robot for some reason takes off at a million mph, and also the
wheel don't rotate anymore on the red robot. Not sure I'm going to be
able to get the red robot's wheels to rotate actually, since it's using
the joint state publisher. We'll see.