Fixed nusim walls and obstacles rotating with robot
After making the change to have my fake obstacle array markers be
relative in position to the robot instead of absolute, a bunch of my
stuff was messed up and I also hadn't checked whether or not rotating
the turblebot would screw any thing up. Spoilers, it did. I spent a
bunch of time implementing transforms for the robot and the obstacles
and laser points and fake obstacles, and now everythign works correctly
even when I'm rotating.