The algorithm is very close, but I have a nan issue
I'm very tired so I'm going to sleep, but the algorithm is in a very
good place I believ and I understand it well. Not sure what's going
on with the nan stuff and I can't get gdb to work for some reason, it
keeps loading me into turtlelib to start the program which prevents me
from putting a breakpoint in the correct place.