Fixed paths and all wheels rotating on robots. Figure dout problem with slam
I figured out the problem with my slam, and it comes from an edge case
where using atan2 gives me something like -3.13 and 3.13, which in
reality are very very close to each other however after subtracting a
negative number from them, it makes them appear very far apart and
my kalman filter tries to correct for this and goes crazy and can't
recover. Still thinking of a way to fix this.